# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Internship Application Tracker API",
    "version": "2.0.0",
    "summary": "REST API for managing an internship application pipeline",
    "description": "Authenticated single-owner API with ownership isolation, follow-up planning, status history, archive/restore, optimistic concurrency, durable reminders, and server-sent notification events. JSON fields use camelCase while MySQL columns use snake_case."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current origin"
    }
  ],
...
  "paths": {
...
    "/applications": {
      "get": {
        "operationId": "listApplications",
        "summary": "List applications",
        "description": "Searches company and position, combines lifecycle and follow-up filters, and paginates on the server.",
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SearchQuery"
          },
          {
            "$ref": "#/components/parameters/StatusQuery"
          },
          {
            "$ref": "#/components/parameters/AttentionQuery"
          },
          {
            "$ref": "#/components/parameters/SortQuery"
          },
          {
            "$ref": "#/components/parameters/DirectionQuery"
          },
          {
            "$ref": "#/components/parameters/ViewQuery"
          },
          {
            "$ref": "#/components/parameters/PageQuery"
          },
          {
            "$ref": "#/components/parameters/LimitQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated applications",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "post": {
        "operationId": "createApplication",
        "summary": "Create an application and its initial status-history entry",
        "tags": [
          "Applications"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationCreateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Application created",
            "headers": {
              "Location": {
                "description": "Relative application URL",
                "schema": {
                  "type": "string",
                  "example": "/api/applications/42"
                }
              },
              "ETag": {
                "$ref": "#/components/headers/VersionEtag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/CsrfInvalid"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ]
      }
    },
...
    "/applications/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ApplicationId"
        }
      ],
      "get": {
        "operationId": "getApplication",
        "summary": "Get one active or archived application",
        "tags": [
          "Applications"
        ],
        "responses": {
          "200": {
            "description": "Application found",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/VersionEtag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidId"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "patch": {
        "operationId": "updateApplication",
        "summary": "Update supplied fields on an active application",
        "description": "Requires the current version through If-Match. A real status change and the application update commit in one transaction.",
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationPatchInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Application updated",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/VersionEtag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/CsrfInvalid"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteApplication",
        "summary": "Permanently delete an archived application",
        "tags": [
          "Applications"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "responses": {
          "204": {
            "description": "Application and status history deleted"
          },
          "400": {
            "$ref": "#/components/responses/InvalidId"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/CsrfInvalid"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ]
      }
    },
...
    "/applications/{id}/archive": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ApplicationId"
        },
        {
          "$ref": "#/components/parameters/IfMatch"
        }
      ],
      "post": {
        "operationId": "archiveApplication",
        "summary": "Archive an active application",
        "tags": [
          "Applications"
        ],
        "responses": {
          "200": {
            "description": "Application archived",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/VersionEtag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/CsrfInvalid"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ]
      }
    },
    "/applications/{id}/restore": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ApplicationId"
        },
        {
          "$ref": "#/components/parameters/IfMatch"
        }
      ],
      "post": {
        "operationId": "restoreApplication",
        "summary": "Restore an archived application",
        "tags": [
          "Applications"
        ],
        "responses": {
          "200": {
            "description": "Application restored",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/VersionEtag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/CsrfInvalid"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "headers": {
      "VersionEtag": {
        "description": "Strong ETag containing the current optimistic-concurrency version",
        "schema": {
          "type": "string",
          "pattern": "^\"[1-9]\\d*\"$",
          "example": "\"3\""
        }
      }
    },
    "parameters": {
      "ApplicationId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Positive unsigned BIGINT as a decimal string",
        "schema": {
          "type": "string",
          "pattern": "^[1-9]\\d*$",
          "maxLength": 20,
          "example": "42"
        }
      },
      "IfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "description": "Quoted application version returned by the previous response",
        "schema": {
          "type": "string",
          "pattern": "^\"[1-9]\\d*\"$",
          "example": "\"3\""
        }
      },
      "SearchQuery": {
        "name": "q",
        "in": "query",
        "description": "Case-insensitive company or position search",
        "schema": {
          "type": "string",
          "maxLength": 120,
          "default": ""
        }
      },
      "StatusQuery": {
        "name": "status",
        "in": "query",
        "description": "Exact status filter",
        "schema": {
          "oneOf": [
            {
              "$ref": "#/components/schemas/ApplicationStatus"
            },
            {
              "type": "string",
              "const": ""
            }
          ],
          "default": ""
        }
      },
      "AttentionQuery": {
        "name": "attention",
        "in": "query",
        "description": "Follow-up date relative to APP_TIMEZONE",
        "schema": {
          "type": "string",
          "enum": [
            "",
            "overdue",
            "today",
            "next7",
            "none"
          ],
          "default": ""
        }
      },
      "SortQuery": {
        "name": "sort",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": [
            "updatedAt",
            "followUpAt",
            "appliedAt",
            "company"
          ],
          "default": "updatedAt"
        }
      },
      "DirectionQuery": {
        "name": "direction",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": [
            "asc",
            "desc"
          ],
          "default": "desc"
        }
      },
      "ViewQuery": {
        "name": "view",
        "in": "query",
        "description": "Active records are the default; archived records are a separate recoverable view",
        "schema": {
          "type": "string",
          "enum": [
            "active",
            "archived"
          ],
          "default": "active"
        }
      },
      "PageQuery": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 1000000,
          "default": 1
        }
      },
      "LimitQuery": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      }
    },
    "schemas": {
      "ApplicationStatus": {
        "type": "string",
        "enum": [
          "wishlist",
          "applied",
          "interview",
          "offer",
          "rejected"
        ]
      },
      "DatabaseTimestamp": {
        "type": "string",
        "pattern": "^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}:\\d{2}\\.\\d{3}$",
        "description": "MySQL TIMESTAMP(3) string in the database session timezone",
        "example": "2026-08-13 09:28:28.123"
      },
      "Application": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "company",
          "position",
          "jobUrl",
          "status",
          "appliedAt",
          "notes",
          "nextAction",
          "followUpAt",
          "createdAt",
          "updatedAt",
          "archivedAt",
          "version"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "42"
          },
          "company": {
            "type": "string",
            "maxLength": 120,
            "example": "VNG Corporation"
          },
          "position": {
            "type": "string",
            "maxLength": 120,
            "example": "Backend Intern"
          },
          "jobUrl": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri",
            "pattern": "^[Hh][Tt][Tt][Pp][Ss]?://",
            "maxLength": 2048,
            "description": "Absolute HTTP or HTTPS URL"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "appliedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "notes": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 5000
          },
          "nextAction": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 240,
            "example": "Email the recruiter with a portfolio link."
          },
          "followUpAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date",
            "example": "2026-08-20"
          },
          "createdAt": {
            "$ref": "#/components/schemas/DatabaseTimestamp"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/DatabaseTimestamp"
          },
          "archivedAt": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DatabaseTimestamp"
              },
              {
                "type": "null"
              }
            ]
          },
          "version": {
            "type": "integer",
            "minimum": 1,
            "maximum": 4294967295
          }
        }
      },
      "ApplicationFields": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "company": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "position": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "jobUrl": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri",
            "pattern": "^[Hh][Tt][Tt][Pp][Ss]?://",
            "maxLength": 2048
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "appliedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "notes": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 5000
          },
          "nextAction": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 240
          },
          "followUpAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          }
        }
      },
      "ApplicationCreateInput": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApplicationFields"
          },
          {
            "required": [
              "company",
              "position"
            ]
          }
        ],
        "example": {
          "company": "VNG Corporation",
          "position": "Backend Intern",
          "status": "applied",
          "nextAction": "Send a portfolio link",
          "followUpAt": "2026-08-20"
        }
      },
      "ApplicationPatchInput": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApplicationFields"
          },
          {
            "minProperties": 1
          }
        ],
        "example": {
          "status": "interview",
          "nextAction": "Prepare system-design examples"
        }
      },
      "Pagination": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ],
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "ApplicationResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Application"
          }
        }
      },
      "ApplicationListResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Application"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
...
      "ErrorDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Request validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InvalidId": {
        "description": "Application id or concurrency header is invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Application not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "Stale version or invalid lifecycle state",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "PreconditionRequired": {
        "description": "If-Match header is required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "JSON request body exceeds 32 KB",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "AuthenticationRequired": {
        "description": "A valid owner session is required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "CsrfInvalid": {
        "description": "The CSRF token is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "RateLimited": {
        "description": "The client exceeded a request rate limit",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tracker_session",
        "description": "Opaque HttpOnly session cookie. Production uses the __Host-tracker_session name."
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Token returned by GET /auth/session or POST /auth/login. Required with the session cookie for state-changing requests."
      }
    }
  }
}